requirements.txt
requests>=2.32.3
tqdm>=4.66
typer>=0.12
python-dotenv>=1.0 